update frontend components UX

+++ frontend/src/components/settings/SettingsComponent.jsx
@@ -1,6 +1,4 @@
 import { useState, useEffect } from 'react';
-import Tooltip from '../../components/tooltip/Tooltip.jsx';
-import 'tippy.js/dist/tippy.css';
 
 const SettingsComponent = () => {
     const [currentPassword, setCurrentPassword] = useState('');
@@ -9,7 +7,9 @@ const SettingsComponent = () => {
     const [username, setUsername] = useState('');
     const [confirmUsername, setConfirmUsername] = useState('');
     const [message, setMessage] = useState('');
+    const [colorMessage , setColorMessage] = useState('');
     const [accountMessage, setAccountMessage] = useState('');
+    const [colorAccountMessage , setColorAccountMessage] = useState('');
     const [userInfo, setUserInfo] = useState(null);
 
     const token = sessionStorage.getItem('jwtToken');
@@ -52,6 +52,7 @@ const SettingsComponent = () => {
             setNewPassword('');
             setConfirmPassword('');
             setMessage('Contraseña cambiada con éxito');
+            setColorMessage('text-green-600');
 
             setTimeout(() => {
                 sessionStorage.removeItem('jwtToken');
@@ -60,6 +61,7 @@ const SettingsComponent = () => {
         } else {
             if(response.status === 400 && data.message === 'Incorrect current password') {
                 setMessage('Contraseña actual incorrecta');
+                setColorMessage('text-red-600');
             }
         }
 
@@ -78,6 +80,7 @@ const SettingsComponent = () => {
             setUsername('');
             setConfirmUsername('');
             setAccountMessage('Cuenta eliminada con éxito');
+            setColorAccountMessage('text-green-600')
 
             setTimeout(() => {
                 sessionStorage.removeItem('jwtToken');
@@ -86,6 +89,7 @@ const SettingsComponent = () => {
         } else {
             if(response.status === 500) {
                 setAccountMessage('Error al eliminar la cuenta');
+                setColorAccountMessage('text-red-600');
             }
         }
 
@@ -93,19 +97,20 @@ const SettingsComponent = () => {
 
 
     const handlePasswordChange = () => {
-        if (newPassword === confirmPassword) {
+        if (newPassword !== '' && confirmPassword !== '' && newPassword === confirmPassword) {
             updatePassword();
         } else {
             setMessage('Las nuevas contraseñas no coinciden');
+            setColorMessage('text-red-600');
         }
     };
 
     const handleAccountDeletion = () => {
-        if (confirmUsername === username) {
+        if (confirmUsername !== '' && username !== '' && confirmUsername === username) {
             deleteUser();
-            setAccountMessage('Cuenta eliminada con éxito');
         } else {
             setAccountMessage('El nombre de usuario no coincide');
+            setColorAccountMessage('text-red-600');
         }
     };
 
@@ -114,7 +119,7 @@ const SettingsComponent = () => {
     }, []);
 
     return (
-        <div className='flex flex-col items-center justify-center min-h-screen bg-gray-700 p-10'>
+        <div className='flex flex-col items-center justify-center min-h-screen bg-gray-700 p-10 rounded-xl'>
             <div className='w-full max-w-2xl p-8 bg-white rounded-lg shadow-lg space-y-10'>
                 <div className='settings-item'>
                         <div className="flex flex-col items-start p-6 rounded-lg bg-orange-50 shadow-md">
@@ -156,48 +161,46 @@ const SettingsComponent = () => {
                                     >
                                         Guardar nueva contraseña
                                     </button>
-                                    {message && <p className="mt-2 text-sm text-green-600">{message}</p>}
+                                    {message && <p className={`mt-2 text-md font-semibold ${colorMessage}`}>{message}</p>}
                                 </div>
                             </div>
                         </div>
                 </div>
                 <div className='settings-item'>
-                    <Tooltip content="Eliminar cuenta">
-                        <div className="flex flex-col items-start p-6 rounded-lg bg-orange-50 shadow-md">
-                            <div className="flex items-center justify-center p-4 rounded-full bg-orange-200 text-white mb-4 shadow-lg">
-                                <span className="material-symbols-outlined text-4xl">
-                                    delete_forever
-                                </span>
-                            </div>
-                            <div className="w-full">
-                                <span className="block text-lg font-semibold text-orange-500 mb-2 text-center">Eliminar cuenta de @{userInfo?.username}</span>
-                                <p className="text-sm text-gray-700 mb-4 text-center">
-                                    Si decides que ya no quieres usar nuestra plataforma, escribe tu nombre de usuario para confirmar y elimina tu cuenta aquí.
-                                </p>
-                                <input
-                                    type="text"
-                                    placeholder="Nombre de usuario"
-                                    value={username}
-                                    onChange={(e) => setUsername(e.target.value)}
-                                    className="mt-1 p-3 block w-full rounded-md border border-gray-300 shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-                                />
-                                <input
-                                    type="text"
-                                    placeholder="Confirma tu nombre de usuario"
-                                    value={confirmUsername}
-                                    onChange={(e) => setConfirmUsername(e.target.value)}
-                                    className="mt-1 p-3 block w-full rounded-md border border-gray-300 shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
-                                />
-                                <button
-                                    onClick={handleAccountDeletion}
-                                    className="mt-4 w-full inline-flex justify-center py-3 px-4 border border-transparent shadow-md text-sm font-medium rounded-md text-white bg-red-600 hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500"
-                                >
-                                    Eliminar cuenta de @{userInfo?.username}
-                                </button>
-                                {accountMessage && <p className="mt-2 text-sm text-green-600">{accountMessage}</p>}
-                            </div>
+                    <div className="flex flex-col items-start p-6 rounded-lg bg-orange-50 shadow-md">
+                        <div className="flex items-center justify-center p-4 rounded-full bg-orange-200 text-white mb-4 shadow-lg">
+                            <span className="material-symbols-outlined text-4xl">
+                                delete_forever
+                            </span>
                         </div>
-                    </Tooltip>
+                        <div className="w-full">
+                            <span className="block text-lg font-semibold text-orange-500 mb-2 text-center">Eliminar cuenta de @{userInfo?.username}</span>
+                            <p className="text-sm text-gray-700 mb-4 text-center">
+                                Si decides que ya no quieres usar nuestra plataforma, escribe tu nombre de usuario para confirmar y elimina tu cuenta aquí.
+                            </p>
+                            <input
+                                type="text"
+                                placeholder="Nombre de usuario"
+                                value={username}
+                                onChange={(e) => setUsername(e.target.value)}
+                                className="mt-1 p-3 block w-full rounded-md border border-gray-300 shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
+                            />
+                            <input
+                                type="text"
+                                placeholder="Confirma tu nombre de usuario"
+                                value={confirmUsername}
+                                onChange={(e) => setConfirmUsername(e.target.value)}
+                                className="mt-1 p-3 block w-full rounded-md border border-gray-300 shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm"
+                            />
+                            <button
+                                onClick={handleAccountDeletion}
+                                className="mt-4 w-full inline-flex justify-center py-3 px-4 border border-transparent shadow-md text-sm font-medium rounded-md text-white bg-red-600 hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500"
+                            >
+                                Eliminar cuenta de @{userInfo?.username}
+                            </button>
+                            {accountMessage && <p className={`mt-2 text-md font-semibold ${colorAccountMessage}`}>{accountMessage}</p>}
+                        </div>
+                    </div>
                 </div>
             </div>
         </div>
